requirements.txt
anyio==4.4.0
certifi==2024.6.2
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openpyxl==3.1.5
progressbar2==4.4.2
python-utils==3.8.2
sniffio==1.3.1
typing_extensions==4.12.2
